pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]


[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.2",
    "pytest-cov>=4.1.0",
    "PyYAML>=6.0",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

